scrapers/requirements.txt
requests==2.32.4
beautifulsoup4==4.13.4
sqlalchemy==2.0.36
psycopg2-binary==2.9.10
python-dotenv==1.0.1
